pyproject.toml
[tool.poetry]
name = "chitalishta-izsledvane-ai-chat"
version = "0.1.0"
description = "RAG system for Bulgarian Chitalishta research data"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic-settings = "^2.6.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
python-docx = "^1.2.0"
chromadb = "^1.0.0"
pybase64 = "^1.3.0"
openai = "^1.0.0"
sentence-transformers = "^2.3.0"
langchain = "^1.2.0"
langchain-openai = "^1.1.0"
langchain-community = "^0.4.1"
langchain-chroma = "^1.1.0"
langchain-huggingface = "^1.2.0"
transformers = "^4.57.3"
requests = "^2.32.0"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
ruff = "^0.14.9"
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

